Extract ProjectsCollectionReference from GroundFirestore
